+++ firstSolar/front/src/ui-components/BeforeHeadCom1440px.jsx
@@ -25,10 +25,10 @@ import { useAccount } from "wagmi";
 export default function BeforeHeadCom1440px(props) {
   const { overrides, ...rest } = props;
   const { isOpen, open, close, setDefaultChain } = useWeb3Modal();
-  const login = useSelector((state) => state.login.login.login);
+  const login = useSelector(state => state.login.login.login);
   const [view, setView] = React.useState(false);
   const dispatch = useDispatch();
-  const connect = useSelector((state) => state.connect.connect.connect);
+  const connect = useSelector(state => state.connect.connect.connect);
   const navigate = useNavigate();
   const { address, isConnecting, isDisconnected } = useAccount();
   const { pathname } = useLocation();
@@ -39,7 +39,7 @@ export default function BeforeHeadCom1440px(props) {
   }, [pathname]);
 
   React.useEffect(() => {
-    const checkIfClickedOutside = (e) => {
+    const checkIfClickedOutside = e => {
       if (view && ref.current && !ref.current.contains(e.target)) {
         setView(false);
       }
